+++ src/gui/dialogs/history_server_dialog.cpp
@@ -140,7 +140,7 @@ void ServerHistoryDialog::refreshInfoFields(int index) {
 
   std::vector< std::vector<Field> > fields = infoFieldsFromType(server_->type());
   std::vector<Field> field = fields[index];
-  for (size_t i = 0; i < field.size(); ++i) {
+  for (int i = 0; i < field.size(); ++i) {
     Field fl = field[i];
     if (fl.isIntegral()) {
       serverInfoFields_->addItem(common::convertFromString<QString>(fl.name), i);
